+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @infinit-xyz/core
 
+## 0.1.0-alpha.5
+
+### Minor Changes
+
+- [#85](https://github.com/infinit-xyz/infinit-library/pull/85)
+  [`128f5d5`](https://github.com/infinit-xyz/infinit-library/commit/128f5d53bb344eb5078436684413ecd12e86a05b) Thanks
+  [@Kittikornl](https://github.com/Kittikornl)! - update to support init-capital module
+
+### Patch Changes
+
+- [#91](https://github.com/infinit-xyz/infinit-library/pull/91)
+  [`51eb6d4`](https://github.com/infinit-xyz/infinit-library/commit/51eb6d4ddc3906459a4f6df23b9b992100331919) Thanks
+  [@melvinmcrn](https://github.com/melvinmcrn)! - bump viem version from 2.21.16 to 2.21.28
+
 ## 0.1.0-alpha.4
 
 ### Minor Changes